add tests/PageObjects/ActionsPage.py
