+++ tests/PageObjects/Actions.py
@@ -1,7 +1,6 @@
 import time
-from abc import abstractmethod
-from selenium.webdriver.common.by import By
-from selenium.common.exceptions import TimeoutException, NoSuchElementException
+
+from selenium.common.exceptions import NoSuchElementException
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.common.action_chains import ActionChains
@@ -12,35 +11,11 @@ from selenium.webdriver.common.keys import Keys
 
 class Actions(object):
 
-    _wait_timeout = 10
     _delay = 0.5
     _a_bit_delay = 0.5
 
-    def __init__(self, driver, server):
+    def __init__(self, driver, server=None):
         self.driver = driver
-        if server:
-            self.load_page(server)
-        self._validate_page()
-
-    @abstractmethod
-    def _validate_page(self):
-        return
-
-    def load_page(self, server, expected_element=(By.TAG_NAME, 'html'),
-                  timeout=_wait_timeout):
-        url = server
-        print("...load_page, url: {u}".format(u=url))
-        try:
-            self.driver.get(url)
-        except TimeoutException:
-            assert(False), "page not found or timeout for {0}".format(url)
-
-        element = EC.presence_of_element_located(expected_element)
-        try:
-            WebDriverWait(self.driver, timeout).until(element)
-        except TimeoutException:
-            assert(False), "page not found or timeout  for {0}".format(url)
-        time.sleep(self._delay)
 
     def find_element_by_id(self, id, descr):
         try:
@@ -170,8 +145,3 @@ class Actions(object):
             print(
                 "NoSuchElementException...could not open vcdat widget")
             raise e
-
-
-class InvalidPageException(Exception):
-    """ Throw this exception when we do not find the correct page """
-    pass
